+++ pepdbagent/utils.py
@@ -1,7 +1,12 @@
 from collections.abc import Iterable
+import json
+from hashlib import md5
+from typing import Tuple, Union
+import ubiquerg
+from .exceptions import RegistryPathError
 
 
-def is_valid_resgistry_path(rpath: str) -> bool:
+def is_valid_registry_path(rpath: str) -> bool:
     """
     Verify that a registry path is valid. Checks for two things:
     1. Contains forward slash ("/"), and
@@ -32,3 +37,53 @@ def all_elements_are_strings(iterable: Iterable) -> bool:
     if not isinstance(iterable, Iterable):
         return False
     return all([isinstance(item, str) for item in iterable])
+
+
+def create_digest(project_dict: dict) -> str:
+    """
+    Create digest for PEP project
+    :param project_dict: project dict
+    :return: digest string
+    """
+    sample_digest = md5(
+        json.dumps(
+            project_dict["_sample_dict"],
+            separators=(",", ":"),
+            ensure_ascii=False,
+            allow_nan=False,
+            sort_keys=True,
+        ).encode("utf-8")
+    ).hexdigest()
+    return sample_digest
+
+
+def registry_path_converter(registry_path: str) -> Tuple[str, str, str]:
+    """
+    Convert registry path to namespace, name, tag
+    :param registry_path: registry path that has structure: "namespace/name:tag"
+    :return: tuple(namespace, name, tag)
+    """
+    if is_valid_registry_path(registry_path):
+        reg = ubiquerg.parse_registry_path(registry_path)
+        namespace = reg["namespace"]
+        name = reg["item"]
+        tag = reg["tag"]
+        return namespace, name, tag
+
+    raise RegistryPathError(f"Error in: '{registry_path}'")
+
+
+def tuple_converter(value: Union[tuple, list, str, None]) -> tuple:
+    """
+    Convert string list or tuple to tuple.
+    # is used to create admin tuple.
+    :param value: Any value that has to be converted to tuple
+    :return: tuple of strings
+    """
+    if isinstance(value, str):
+        value = [value]
+    if value:
+        return tuple(value)
+    return tuple(
+        " ",
+    )